feat: add location id prop to error boundary component
This allows the use of the error boundary component in more places than one. For instance the boundary could exist to catch errors within the main container, so that at least the header and footer are preserved. And then of course there is a top level component to catch any error.