test(browser): various improvements to homepage test
I moved away from wanting *every* test to be run in both desktop and mobile view. Now the only conditional viewport tests are for the aspects that are actually different based on the viewport size, which is really just links vs. hamburger menu in the header.

In the future, the totality of the app on a small screen size would be properly tested by using Appium or something to test in a true mobile device setting.